* merge impl basic

* comments

+++ include/son8/sortable/util/range.hxx
@@ -2,6 +2,7 @@
 #define SON8_SORTABLE_UTIL_RANGE_HXX
 
 #include <cassert>
+#include <cstddef>
 
 namespace son8::sortable {
     // Range class represents a half-open range [beg, end)
@@ -18,7 +19,7 @@ namespace son8::sortable {
         Type *beg( ) const { return beg_; }
         Type *end( ) const { return end_; }
         Type *mid( ) const { return beg_ + ( ( end_ - beg_ ) >> 1 ); }
-        auto size( ) const { return end_ - beg_; }
+        auto size( ) const { return static_cast< std::size_t >( end_ - beg_ ); }
     }; // class Range
 
     template< typename Type >
